+++ deser-json/src/lib.rs
@@ -8,6 +8,10 @@
 //! let vec: Vec<u64> = deser_json::from_str("[1, 2, 3, 4]").unwrap();
 //! let json = deser_json::to_string(&vec).unwrap();
 //! ```
+//!
+//! By default this crate has no dependency crates other than `deser`, but optionally
+//! the `speedups` feature can be enabled in which case this also uses `ryu` and `itoa`
+//! crates are used for number formatting.
 mod de;
 mod ser;